Always use composer for running phpunit.
Originally, phpunit installation in Moodle was only supported via pear,
but that changed in Moodle 2.4, and was backported to 2.3.3.

The code was mistakenly checking for a minimum version of 25, which is
not correct as composer installation is preffered for all versions of
Moodle 2.4.

There is an argument to keep non-composer compatibility for 2.3 versions
< 2.3.3, but I selfishly would prefer we switched completely, because I
don't want to install pear for my 2.3 work.