Adding support for subsystems in PluginManager
Yup, it should not really have been mixed up with plugins,
but I am sort of lazy to write 'component' library so
it will be there for now, and I am sure we will live with it.

Also, why the hell did I make that class with only static
methods? I must have had a reason at the time...