Add support for optional db.namePrefix setting
Currently, the default database name is the same as the Moodle instance's name.
This is suboptimal for development setups where the database engine is used for
other projects as well (Moodle developer can easily have Mahara or Mediawiki
databases present as well, for example).

I believe it's a good idea to make it clear that certain database is under
mdk's control. Having a prefix (such as `mdk_`) seems to work well.