Add basic support for peer review transitions
These will need to be tweaked a bit, but currently offer:
* start peer review; and
* fail peer review.

This change does not incorporate an --end-review, or -pass-review phase.
The end-review should be a meta for comment as per our new PR guideliens
The pass-review should be identical to the fail review, except that it
should transition to "Waiting for integration review" if possible.