Fix precheck command after phpcs renaming and add missing checks.
In https://github.com/moodlehq/moodle-local_ci/pull/70 the php code
sniffer abbreviation was changed to phpcs. In various other isues we've
added more CI checks which are missing from mdk.

Ideally this should be rewritten to output the result of all the checks,
regardless of whether they are defined in mdk, but thats for another
day.