Clean up `Pull X.Y` fields belonging to legacy branches
As part of the Moodle release process, custom fields belonging to legacy
branches are being removed from the tracker. As of the 4.4 release
`Pull X.Y` custom fields belonging to versions 4.0 and lower are no
longer in the tracker. So we might as well clean up config-dist.json
of the custom fields belonging to these legacy branches.