Update gt to link to the moodle root
I am informed that some people create their own symlink for Moodle to
link straight to the `public` directory. As a result though the `gt`
command points to the wrong place.

Rather than having `gt` point to the public directory, this changes it
to point to the storage directory behind the normal symlink.